Generate libs for wrapper at other corners the macro is characterized at
In #11, I point out that the "characterization" for the wrapper is a simple modification of the canonical module's lib file.

This commit creates a script to apply a similar transformation to the other corner `.lib` files and includes the result of running that script. The output for `…wrapper_tt…` is identical to its previous state.